% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_questionnaire_duplicates.R
\name{check_questionnaire_duplicates}
\alias{check_questionnaire_duplicates}
\title{Check for Duplicate Questionnaire Events}
\usage{
check_questionnaire_duplicates(responses_df, program_name, milestone_name)
}
\arguments{
\item{responses_df}{data.frame; Responses from get_responses_df()}

\item{program_name}{character; Program type}

\item{milestone_name}{character; Milestone}
}
\value{
list with diagnostic information
}
\description{
Diagnostic function to verify no duplicate questionnaire events
exist in a specific context after data transformation
}
