% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators_df.R
\name{get_indicators_df}
\alias{get_indicators_df}
\title{Get Indicators Data Frame}
\usage{
get_indicators_df()
}
\value{
a data frame of questionnaire events with indicator scores
}
\description{
Get questionnaire events with the indicator metric scores
calculated.
}
