% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions_df.R
\name{get_questions_df}
\alias{get_questions_df}
\title{Get Questions Data Frame}
\usage{
get_questions_df()
}
\value{
data frame of questions
}
\description{
Gets a data frame of question numbers and question text.
}
