% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator_alluvial.R
\name{plot_indicator_alluvial}
\alias{plot_indicator_alluvial}
\title{Plot Alluvial Diagram for Indicator Questions}
\usage{
plot_indicator_alluvial(indicator_name)
}
\arguments{
\item{indicator_name}{Character; one of the indicator names}
}
\value{
A ggplot2 alluvial plot object
}
\description{
Shows flow from individual question responses to indicator score
}
