% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator_by_div.R
\name{plot_indicator_by_div}
\alias{plot_indicator_by_div}
\title{Plot Indicator by USACE Division}
\usage{
plot_indicator_by_div(indicators_df, indicator_name)
}
\arguments{
\item{indicators_df}{data.frame; A data frame of questionnaire events
with indicator scores returned by
\code{get_indicators_df}.}

\item{indicator_name}{character; The indicator name to be ploted. One of
\code{levels(get_responses_df()$INDICATOR)}:
"Confidence", "Cost", "QA", "QC", "Schedule",
"Scope", "Team".}
}
\value{
a ggplot2 object
}
\description{
Plot indicator by USACE Division.
}
