% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator_prog_mile.R
\name{plot_indicator_prog_mile}
\alias{plot_indicator_prog_mile}
\title{Plot Alluvial Diagrams for an Indicator by Milestones}
\usage{
plot_indicator_prog_mile(indicator_name, program_name)
}
\arguments{
\item{indicator_name}{character; Questionnaire indicator name. One of:
\code{levels(get_indicators_df()$INDICATOR)}}

\item{program_name}{character; Program name. One of:
\code{levels(get_indicators_df()$PROGRAMTYPE_NAME)}}
}
\value{
a ggplot2 object
}
\description{
Plots a set of Alluvial Diagrams for all project milestones.
}
