% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_question_by_design_team.R
\name{plot_question_by_design_team}
\alias{plot_question_by_design_team}
\title{Plot Question by Design Team}
\usage{
plot_question_by_design_team(questions_df, responses_df, question_number)
}
\arguments{
\item{questions_df}{data.frame; A data frame of questions
returned by \code{get_questions_df}.}

\item{responses_df}{data.frame; A data frame of responses
returned by \code{get_responses_df}.}

\item{question_number}{integer; The question number to be ploted.}
}
\value{
a ggplot2 object
}
\description{
Plot question by Design Team.
or Military).
}
