% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_question_by_div.R
\name{plot_question_by_div}
\alias{plot_question_by_div}
\title{Plot Question by USACE Divisions}
\usage{
plot_question_by_div(questions_df, responses_df, question_number)
}
\arguments{
\item{questions_df}{data.frame; A data frame of questions
returned by \code{get_questions_df}.}

\item{responses_df}{data.frame; A data frame of responses
returned by \code{get_responses_df}.}

\item{question_number}{integer; The question number to be ploted.}
}
\value{
a ggplot2 object
}
\description{
Plot question by USACE Divisions.
}
