% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_indicator.R
\name{render_indicator}
\alias{render_indicator}
\title{Render Indicator}
\usage{
render_indicator(indicator_df, indicator_name)
}
\arguments{
\item{indicator_df}{data.frame; A data frame of questionnaire events
with indicator scores returned by
\code{get_indicator_df}.}

\item{indicator_name}{character; The indicator name to be ploted. One of
\code{levels(get_responses_df()$INDICATOR)}:
"Confidence", "Cost", "QA", "QC", "Schedule",
"Scope", "Team".}
}
\value{
a markdown string representing the output for each indicator
}
\description{
Render the markdown output for an indicator.
}
