% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_question.R
\name{render_question}
\alias{render_question}
\title{Render Question}
\usage{
render_question(questions_df, responses_df, question_number)
}
\arguments{
\item{questions_df}{data.frame; A data frame of questions
returned by \code{get_questions_df}.}

\item{responses_df}{data.frame; A data frame of responses
returned by \code{get_responses_df}.}

\item{question_number}{integer; The question number to be ploted.}
}
\value{
a markdown string representing the output for each question
}
\description{
Render the markdown output for a question.
}
