% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alluvial_df.R
\name{get_alluvial_df}
\alias{get_alluvial_df}
\title{Get Alluvial Data for Question-Indicator Flow}
\usage{
get_alluvial_df()
}
\value{
A data frame formatted for alluvial plotting
}
\description{
Prepare data showing how question responses flow to indicator scores
}
